\name{brp}
\alias{brp}
\alias{EU_data}
\alias{USdata1}
\alias{USdata2}
\alias{coordUS}
\alias{alcohol}
\alias{gdp}
\alias{gini}
\alias{dat}
\alias{pop}
\alias{popEU}
\alias{debtEU}
\alias{coordsEU}
\title{data sets for testing the sparkTable package}
\usage{
data(brp)
data(EU_data)
data(USdata1)
data(USdata2)
data(coordUS)
data(alcohol)
data(gdp)
data(gini)
data(dat)
data(pop)
data(popEU)
data(debtEU)
data(coordsEU)
}
\description{
  brp...Gross Regional Product Austria
  gdp...GDP for 41 different regions and from 1997 to 2008
  gini...gini Index for different european countries from 2004 to 2007 
  dat...Production Index in Austria 
  pop...the number of people living in Austria in different subgroups and years
  popEU ... population of the EU 
  debtEU ... debt of the European countries 
  coordsEU ... coordinates of the European capitals
  alcohol ... alcohol conumption in the European Union from 1970 until 2008
  EU_data ... Unemployment rates of females in the EU from 1997 until 2008
  USdata1 ... Health incurance coverage from US states in 2009
  USdata2 ... Health incurance by goverment and private insurance from US states in 2009
  coordUS ... coordinates of the US states (centres)
}
\examples{
data(brp)
data(gdp)
data(gini)
data(pi)
data(population)
}
\author{Bernhard Meindl, Alexander Kowarik, Statistics Austria  }
