\name{brp}
\alias{brp}
\alias{gdp}
\alias{gini}
\alias{dat}
\alias{pop}
\title{data sets for testing the sparkTable package}
\usage{
data(brp)
data(gdp)
data(gini)
data(dat)
data(pop)
}
\description{
  brp...Gross Regional Product Austria
  gdp...GDP for 41 different regions and from 1997 to 2008
  gini...gini Index for different european countries from 2004 to 2007 
  dat...Production Index in Austria 
  pop...the number of people living in Austria in different subgroups and years
}
\examples{
data(brp)
data(gdp)
data(gini)
data(pi)
data(population)
}
\author{Bernhard Meindl, Alexander Kowarik, Statistics Austria  }
